Move logic for if a post should be pushed
References #328
